+++ crates/tests/src/main.rs
@@ -1,5 +1,5 @@
 use clap::{App, Arg, SubCommand};
-use gw_tests::system_tests::{bad_block, bad_challenge, test_mode_control};
+use gw_tests::system_tests::{bad_block, bad_challenge, utils};
 use std::path::Path;
 
 fn main() -> Result<(), String> {
@@ -16,16 +16,48 @@ fn main() -> Result<(), String> {
                         .help("Get global state"),
                 )
                 .arg(
-                    Arg::with_name("test-blocks")
-                        .long("test-blocks")
-                        .short("t")
+                    Arg::with_name("issue-blocks")
+                        .long("issule-blocks")
+                        .short("i")
                         .takes_value(true)
-                        .help("Issue test blocks"),
+                        .help("Issue empty blocks"),
                 ),
         )
         .subcommand(
             SubCommand::with_name("bad-block")
                 .about("Issue bad block")
+                .arg(
+                    Arg::with_name("from-privkey-path")
+                        .short("f")
+                        .takes_value(true)
+                        .required(true)
+                        .help("from user privkey path"),
+                )
+                .arg(
+                    Arg::with_name("to-privkey-path")
+                        .short("t")
+                        .takes_value(true)
+                        .required(true)
+                        .help("To user privkey path"),
+                )
+                .arg(
+                    Arg::with_name("scripts-deploy-result-path")
+                        .short("d")
+                        .takes_value(true)
+                        .required(true)
+                        .help("Scripts deploy result file path"),
+                )
+                .arg(
+                    Arg::with_name("config-file-path")
+                        .short("c")
+                        .takes_value(true)
+                        .required(true)
+                        .help("godwoken node config file path"),
+                ),
+        )
+        .subcommand(
+            SubCommand::with_name("package-tx")
+                .about("Issue a test block containing tx")
                 .arg(
                     Arg::with_name("from-privkey-path")
                         .short("f")
@@ -69,14 +101,14 @@ fn main() -> Result<(), String> {
     match matches.subcommand() {
         ("test-mode", Some(m)) => {
             if m.is_present("global-state") {
-                let state = test_mode_control::get_global_state()?;
+                let state = utils::get_global_state()?;
                 println!("global state is: {:?}", state);
-            } else if m.is_present("test-blocks") {
+            } else if m.is_present("issue-blocks") {
                 let count = m
-                    .value_of("test-blocks")
+                    .value_of("issue-blocks")
                     .map(|c| c.parse().expect("count of blocks"))
                     .unwrap();
-                test_mode_control::issue_test_blocks(count)?;
+                utils::issue_blocks(count)?;
             } else {
                 app.print_help().expect("print help");
             }
@@ -94,6 +126,22 @@ fn main() -> Result<(), String> {
                 deployment_path,
             ) {
                 log::error!("Issue bad block error: {}", err);
+                std::process::exit(-1);
+            }
+        }
+        ("package-tx", Some(m)) => {
+            let deployment_path = Path::new(m.value_of("scripts-deploy-result-path").unwrap());
+            let from_privkey_path = Path::new(m.value_of("from-privkey-path").unwrap());
+            let to_privkey_path = Path::new(m.value_of("to-privkey-path").unwrap());
+            let config_path = Path::new(m.value_of("config-file-path").unwrap());
+
+            if let Err(err) = utils::package_a_transaction(
+                from_privkey_path,
+                to_privkey_path,
+                config_path,
+                deployment_path,
+            ) {
+                log::error!("Package a transaction error: {}", err);
                 std::process::exit(-1);
             }
         }
