+++ crates/tests/src/system_tests/mod.rs
@@ -1,3 +1,3 @@
 pub mod bad_block;
 pub mod bad_challenge;
-pub mod test_mode_control;
+pub mod utils;
